Moving full database tests to separate file
