+++ tests/conftest.py
@@ -28,16 +28,16 @@ def db():
     db = Database(connection_string, reference_tables=REFERENCE_TABLES)
 
     # The input data is NOT correct; that needs to be fixed or this commented out
-    # # Load data into an in-memory sqlite database first, for performance
-    # db = Database(
-    #     "sqlite://", reference_tables=REFERENCE_TABLES
-    # )  # creates and connects to a temporary in-memory database
-    # db.load_database(
-    #     DB_PATH, verbose=False
-    # )  # loads the data from the data files into the database
-    # db.dump_sqlite(DB_NAME)  # dump in-memory database to file
-    # db = Database(
-    #     "sqlite:///" + DB_NAME, reference_tables=REFERENCE_TABLES
-    # )  # replace database object with new file version
+    # Load data into an in-memory sqlite database first, for performance
+    db = Database(
+        "sqlite://", reference_tables=REFERENCE_TABLES
+    )  # creates and connects to a temporary in-memory database
+    db.load_database(
+        DB_PATH, verbose=False
+    )  # loads the data from the data files into the database
+    db.dump_sqlite(DB_NAME)  # dump in-memory database to file
+    db = Database(
+        "sqlite:///" + DB_NAME, reference_tables=REFERENCE_TABLES
+    )  # replace database object with new file version
 
     return db
